Extract trajectory evaluator.
