+++ pyproject.toml
@@ -90,10 +90,11 @@ ml = [
 
 [project.scripts]
 flatland-demo = "flatland.cli:demo"
+# deprecated, use flatland- prefix for all scripts
 evaluator = "flatland.cli:evaluator"
 flatland-evaluator = "flatland.cli:evaluator"
 flatland-trajectory-generate-from-policy = "flatland.trajectories.trajectories:generate_trajectory_from_policy"
-flatland-trajectory-evaluate = "flatland.trajectories.trajectories:evaluate_trajectory"
+flatland-trajectory-evaluate = "flatland.evaluators.trajectory_evaluator:evaluate_trajectory"
 
 [tool.setuptools.packages]
 find = { }
